Change type of HealthBarChunks in Constants

+++ src/Constants.hpp
@@ -101,6 +101,6 @@ const double ConfettiMinGravity = 0.7;
 const double ConfettiMaxGravity = 1.4;
 
 //Health Bar
-const int HealthBarChunks = 10;
+const double HealthBarChunks = 10;
 
 #endif
